renames communication package
