+++ communication/communication-key-value/src/main/java/jakarta/nosql/kv/KeyValueQueryParser.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package jakarta.nosql.key;
+package jakarta.nosql.kv;
 
 import jakarta.nosql.QueryException;
 import jakarta.nosql.Value;
