+++ mapping/mapping-key-value/src/main/java/jakarta/nosql/mapping/key/KeyValueWorkflow.java
@@ -16,7 +16,7 @@
 package jakarta.nosql.mapping.key;
 
 
-import jakarta.nosql.key.KeyValueEntity;
+import jakarta.nosql.kv.KeyValueEntity;
 
 import java.util.function.UnaryOperator;
